@@ -20,7 +20,7 @@
         "exclude": ["/docs",  "/docs/build"]
     },
     "require": {
-        "php": "^7.3"
+        "php": "^7.4"
     },
     "autoload": {
         "classmap": [
