@@ -81,7 +81,7 @@ TABLE1;
      * @return string Html table with all data for a method
      */
     private function renderAction(string $content, string $headerData = ''): string {
-        $methodKey = $this->rep->activeMethod[0];
+        $methodKey = $this->rep->currentAction[0];
         $headerCol3 = $this->getMethodName() . $headerData;
         $sign = ($this->signs[$methodKey]) ?? '';
         $color = substr($this->colors[$methodKey], 1);
@@ -107,7 +107,7 @@ TABLE1;
      * @return string Output generated by methods to be called
      */
     public function magic(): string {
-        $method = $this->rep->activeMethod[0];
+        $method = $this->rep->currentAction[0];
         switch ($method) {
             case 'groupHeader':
             case 'groupFooter':
@@ -139,14 +139,14 @@ TABLE1;
      * is returned.
      */
     private function getMethodName(): string {
-        if (isset($this->rep->activeMethod[3])) {
+        if (isset($this->rep->currentAction[3])) {
             // direct prototype call from report class
-            [$key, $typ, $action, $ownerAction, $ownerTyp] = $this->rep->activeMethod;
+            [$key, $typ, $action, $ownerAction, $ownerTyp] = $this->rep->currentAction;
             $action = $ownerAction;
             $typ = $ownerTyp;
         } else {
             // prototype call from owner class via $report->prototype()
-            [$key, $typ, $action] = $this->rep->activeMethod;
+            [$key, $typ, $action] = $this->rep->currentAction;
         }
         switch ($typ) {
             // For typ method return only the method name (second array element)
