@@ -121,7 +121,7 @@ class ObjectDataHandlerTest extends TestCase {
      * @dataProvider dataSources
      */
     public function testGetDimData($source, $expected, ...$params) {
-         $dim = new Dimension(1, 'object', $source, 'Testclass', null, null, null, ...$params);
+         $dim = new Dimension(1, 'object', $source, 'Testclass', null, ...$params);
           $this->stack = $dim->dataHandler;
         $mp = Factory::properties();
         $total = new Collector();
